move db connection to new file
